backend/requirements.txt
fastapi
uvicorn[standard]
pydantic-settings
supabase
langchain-core
langgraph
google-generativeai
python-dotenv
httpx
python-jose[cryptography]
PyJWT
pytest
slowapi
